third_party/blink/web_tests/external/wpt/wasm/jsapi/table/set-reftypes.tentative.any-expected.txt

This is a testharness.js-based test.
FAIL Arguments for anyfunc table set assert_equals: expected (function) function "function 0() { [native code] }" but got (object) null
FAIL Arguments for externref table set WebAssembly.Table(): Descriptor property 'element' must be a WebAssembly reference type
Harness: the test ran to completion.
